+++ AssetEditor/Properties/launchSettings.json
@@ -9,6 +9,11 @@
       "commandLineArgs": "-devcfg Kitbash_Karl",
       "nativeDebugging": false
     },
+    "AE-MeshFitter": {
+      "commandName": "Project",
+      "commandLineArgs": "-devcfg Kitbash_MeshFitter",
+      "nativeDebugging": false
+    },
     "AE-AnimPack_Wh3": {
       "commandName": "Project",
       "commandLineArgs": "-devcfg AnimPack_WH3",
