+++ Editors/Kitbashing/KitbasherEditor/ChildEditors/MeshFitter/MeshFitterWindow.xaml
@@ -1,4 +1,5 @@
-﻿<windowhandling:AssetEditorWindow  xmlns:windowhandling="clr-namespace:WindowHandling;assembly=Shared.Ui" x:Class="Editors.KitbasherEditor.ChildEditors.MeshFitter.MeshFitterWindow"
+﻿<windowhandling:AssetEditorWindow  xmlns:windowhandling="clr-namespace:WindowHandling;assembly=Shared.Ui"
+                                   x:Class="Editors.KitbasherEditor.ChildEditors.MeshFitter.MeshFitterWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
